# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountCreated": {
        "additionalProperties": false,
        "properties": {
          "controller_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Controller Id"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "recovery_key": {
            "title": "Recovery Key",
            "type": "string"
          }
        },
        "required": [
          "email",
          "recovery_key"
        ],
        "title": "AccountCreated",
        "type": "object"
      },
...
      "EmailLinkProof": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "pattern": "^[0-9a-f]{32}\\.[A-Za-z0-9_-]{43}$",
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "EmailLinkProof",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvitationAccepted": {
        "additionalProperties": false,
        "properties": {
          "organization_id": {
            "format": "uuid",
            "title": "Organization Id",
            "type": "string"
          }
        },
        "required": [
          "organization_id"
        ],
        "title": "InvitationAccepted",
        "type": "object"
      },
      "InvitationCreate": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "role": {
            "$ref": "#/components/schemas/OrganizationRole",
            "default": "viewer"
          },
          "site_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 100,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Ids"
          }
        },
        "required": [
          "email"
        ],
        "title": "InvitationCreate",
        "type": "object"
      },
      "InvitationPreview": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/OrganizationRole"
          }
        },
        "required": [
          "email",
          "organization_name",
          "role"
        ],
        "title": "InvitationPreview",
        "type": "object"
      },
      "InvitationRead": {
        "additionalProperties": false,
        "properties": {
          "access_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Expires At"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "role": {
            "$ref": "#/components/schemas/OrganizationRole"
          },
          "site_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Ids"
          },
          "used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used At"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "site_ids",
          "access_expires_at",
          "expires_at",
          "used_at",
          "revoked_at"
        ],
        "title": "InvitationRead",
        "type": "object"
      },
...
      "OrganizationRole": {
        "description": "Роль користувача всередині конкретної організації.",
        "enum": [
          "owner",
          "admin",
          "operator",
          "viewer",
          "service"
        ],
        "title": "OrganizationRole",
        "type": "string"
      },
...
      "RegistrationComplete": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Display Name",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "token": {
            "pattern": "^[0-9a-f]{32}\\.[A-Za-z0-9_-]{43}$",
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "display_name",
          "password"
        ],
        "title": "RegistrationComplete",
        "type": "object"
      },
      "RegistrationPreview": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "RegistrationPreview",
        "type": "object"
      },
      "RegistrationStart": {
        "additionalProperties": false,
        "properties": {
          "controller_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Controller Id"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "RegistrationStart",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend API платформи TechBaza IoT Pump Control",
    "title": "TechBaza Backend",
    "version": "0.49.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/registration/complete": {
      "post": {
        "operationId": "register_complete_api_v1_auth_registration_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationComplete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Complete",
        "tags": [
          "personal-accounts"
        ]
      }
    },
    "/api/v1/auth/registration/inspect": {
      "post": {
        "operationId": "register_inspect_api_v1_auth_registration_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLinkProof"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Inspect",
        "tags": [
          "personal-accounts"
        ]
      }
    },
    "/api/v1/auth/registration/start": {
      "post": {
        "operationId": "register_start_api_v1_auth_registration_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Start",
        "tags": [
          "personal-accounts"
        ]
      }
    },
...
    "/api/v1/connect/{controller_id}/account-registration": {
      "post": {
        "operationId": "label_register_start_api_v1_connect__controller_id__account_registration_post",
        "parameters": [
          {
            "in": "path",
            "name": "controller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Controller Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Label Register Start",
        "tags": [
          "personal-accounts"
        ]
      }
    },
...
    "/api/v1/invitations/accept": {
      "post": {
        "operationId": "invitation_accept_api_v1_invitations_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLinkProof"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invitation Accept",
        "tags": [
          "personal-accounts"
        ]
      }
    },
    "/api/v1/invitations/inspect": {
      "post": {
        "operationId": "invitation_inspect_api_v1_invitations_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLinkProof"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invitation Inspect",
        "tags": [
          "personal-accounts"
        ]
      }
    },
    "/api/v1/invitations/register": {
      "post": {
        "operationId": "invitation_register_api_v1_invitations_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationComplete"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Invitation Register",
        "tags": [
          "personal-accounts"
        ]
      }
    },
...
    "/api/v1/organizations/{organization_id}/invitations": {
      "get": {
        "operationId": "invitation_list_api_v1_organizations__organization_id__invitations_get",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvitationRead"
                  },
                  "title": "Response Invitation List Api V1 Organizations  Organization Id  Invitations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invitation List",
        "tags": [
          "personal-accounts"
        ]
      },
      "post": {
        "operationId": "invitation_create_api_v1_organizations__organization_id__invitations_post",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invitation Create",
        "tags": [
          "personal-accounts"
        ]
      }
    },
    "/api/v1/organizations/{organization_id}/invitations/{invitation_id}": {
      "delete": {
        "operationId": "invitation_revoke_api_v1_organizations__organization_id__invitations__invitation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "organization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Invitation Revoke",
        "tags": [
          "personal-accounts"
        ]
      }
    },
...
  }
}